% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_spots.R
\name{enhance_spots}
\alias{enhance_spots}
\title{Enhance Spots}
\usage{
enhance_spots(
  dirpath,
  image_filename,
  image_folder_name = "ImageResults",
  method = "wth"
)
}
\arguments{
\item{dirpath}{A character vector indicating the dirpath for the
Pamstation data folder containing the image data.}

\item{image_filename}{A character vector indicating the
filename of the *.tiff array image file to be processed,
contained within the Pamstation data folder specified in
the \strong{dirpath} argument, within the \strong{image_folder_name}
subdirectory.}

\item{image_folder_name}{A character vector indicating the
name of the ImageData subfolder contained within the Pamstation
data folder specified in \strong{dirpath} - default is "ImageResults".}

\item{method}{A character vector indicating the
method to use for image enhancement: must be in
c("wth", "dog"), for either the "White Top-Hat
Transformation" or the "Difference of Gaussians"
methods respectively.}
}
\value{
A named list, with one list item being named "enhanced_array",
which contains an S4Vectors Dataframe of single-channel greyscale intensities
enhanced for visibility
}
\description{
This function ingests an image file in tiff format
(typically in an \code{ImageResults} subfolder) from
the Pamgene Pamstation kinome array device
and enhances the image so that spots are more easily discerned.
}
\examples{
test_dir <- system.file('extdata', 'test_dir', package = "pamstationR")
enhance_spots(dirpath = test_dir, image_filename='test_image_PTK.tif')
}
\keyword{data}
\keyword{image}
\keyword{image-analysis}
\keyword{import}
\keyword{preprocessing}
