% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_folder.R
\name{import_data_folder}
\alias{import_data_folder}
\title{Import Pamstation folder}
\usage{
import_data_folder(
  dirpath,
  array_annotation,
  sample_annotation,
  image_folder_name = "ImageResults"
)
}
\arguments{
\item{dirpath}{A character vector indicating the dirpath for the
Pamstation data folder containing the image data.}

\item{image_folder_name}{A character vector indicating the
name of the ImageData subfolder contained within the Pamstation
data folder specified in \strong{dirpath} - default is "ImageResults".}
}
\value{
A SummarizedExperiment object representing the intensity
statistics for each probe, well, and chip.
}
\description{
This function imports a data folder, containing image data in tiff format
(typically in an \code{ImageResults} subfolder) from
the Pamgene Pamstation kinome array device.
}
\examples{
# Example 1: A data folder at `/Users/user/project/data/pamstation_output_folder/`
import_data_folder(dirpath = "/Users/user/project/data/pamstation_output_folder/",
image_folder_name = "ImageResults")

# Example 2: A data folder within a data folder inside of the current working directory, at `./data/pamstation_output_folder/` where the image_folder_name is "ImageResults" (default)
import_data_folder(dirpath = "data/pamstation_output_folder/")
}
\seealso{
\code{\link[stats]{setNames}}, \code{\link{get_bmi_category}}
(assuming another function exists)
}
\keyword{data}
\keyword{image}
\keyword{import}
