% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tiff.R
\name{read_tiff}
\alias{read_tiff}
\title{Read TIFF file}
\usage{
read_tiff(dirpath, image_filename, image_folder_name = "ImageResults")
}
\arguments{
\item{dirpath}{A character vector indicating the dirpath for the
Pamstation data folder containing the image data.}

\item{image_filename}{A character vector indicating the
filename of the *.tiff array image file to be processed,
contained within the Pamstation data folder specified in
the \strong{dirpath} argument, within the \strong{image_folder_name}
subdirectory.}

\item{image_folder_name}{A character vector indicating the
name of the ImageData subfolder contained within the Pamstation
data folder specified in \strong{dirpath} - default is "ImageResults".}
}
\value{
A named list, with one list item being named "raw_array",
which contains an S4Vectors Dataframe of single-channel greyscale intensities
enhanced for visibility
}
\description{
This function ingests an image file in tiff format
(typically in an \code{ImageResults} subfolder) from
the Pamgene Pamstation kinome array device
}
\examples{
test_dir <- system.file('extdata', 'test_dir', package = "pamstationR")
read_tiff(dirpath = test_dir, image_filename='test_image_PTK.tif')
}
\keyword{data}
\keyword{image}
\keyword{import}
\keyword{raw-data}
