% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_folder.R
\name{import_data_folder}
\alias{import_data_folder}
\title{Import Pamstation folder}
\usage{
import_data_folder(
  dirpath,
  array_annotation,
  sample_annotation,
  image_folder_name = "ImageResults"
)
}
\arguments{
\item{dirpath}{A character vector indicating the dirpath for the
Pamstation data folder containing the image data.}

\item{array_annotation}{A character vector indicating the
name of the array annotation file contained within the Pamstation
data folder specified in \strong{dirpath}.
This should contain annotations for every spot on the array and
their corresponding peptide/sequence/etc. annotations.}

\item{sample_annotation}{A character vector indicating the
name of the sample annotation file contained within the Pamstation
data folder specified in \strong{dirpath}.
This should contain annotations for every sample on the array and
(one per well) with any relevant information included.}

\item{image_folder_name}{A character vector indicating the
name of the ImageData subfolder contained within the Pamstation
data folder specified in \strong{dirpath} - default is "ImageResults".}
}
\value{
A SummarizedExperiment object representing the intensity
statistics for each probe, well, and chip.
}
\description{
This function imports a data folder, containing image data in tiff format
(typically in an \code{ImageResults} subfolder) from
the Pamgene Pamstation kinome array device.
}
\examples{
temp_dir <- withr::local_tempdir('folder_that_exists')
inner_name <- "ImageResults"
dir.create(file.path(temp_dir, inner_name))
array_annotation_file <- 'array_annotation.txt'
file.create(file.path(temp_dir, array_annotation_file))
sample_annotation_file <- 'sample_annotation.txt'
file.create(file.path(temp_dir, sample_annotation_file))
import_data_folder(dirpath = temp_dir,
array_annotation = file.path(temp_dir, array_annotation_file),
sample_annotation = file.path(temp_dir, sample_annotation_file))
}
\keyword{data}
\keyword{image}
\keyword{import}
